Directory Contents
Will list all contents of the provided directory. May not see everything in the directory (during troubleshooting it turned out that - at least on Windows 10 - some files will be opted out for an unknown reason) but will output the result in a result.txt file created in the directory from where the program was initiated.