Down Detector
Requires installation of Requests (https://2.python-requests.org/en/master/) library to work. Details regarding this are in the code body. Will run a simple GET request to the specified target; if it has a self-signed SSL certificate, Down Detector will not verify it. To help the user realize if the problem lies on the server's side or theirs.