Backup manager
Will create a .tar archive of the folder or file a user specifies. Script interacts with the user.

If an absolute path is provided, tar will include every folder in the archive, even though not entirely. For example, if the path is /home/user/Desktop/directroy/, the archive will contain folders home, user, Desktop, and directory. The last one will have its contents while all others will have only subsequent folders specified in the path as contents.
If a relative path is provided, the result depends on where the script file is located.

As a rule of thumb, it is better to run the script where the target is placed and provide just names of the folders. However, because Linux mv command is used, both relative and absolute paths can be specified when asked where the backups should reside.